improved error handling
moved getWeather function into a separate module

minor changes
