+++ src/index.html
@@ -99,7 +99,7 @@
             <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
           </div>
           <div class="modal-body">
-            <form action="submit" class="change-location-form">
+            <form class="change-location-form">
               <div class="input-group mb-3">
                 <input type="text" class="form-control form-city" placeholder="Enter a city" required>
               </div>
@@ -122,6 +122,6 @@
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-u1OknCvxWvY5kfmNBILK2hRnQC3Pr17a+RTT6rIHI7NnikvbZlHgTPOOmMi466C8" crossorigin="anonymous"></script>
   <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-  <script type="module" src="../js/app.js"></script>
+  <script type="module" src="./js/app.js"></script>
 </body>
 </html>
