+++ src/style.css
@@ -26,7 +26,6 @@ Main Card
 
 .main-card {
 
-  /* From https://css.glass */
   background: rgba(255, 255, 255, 0.38);
   border-radius: 16px;
   box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
